requirements.txt
transformers==4.57.5
gradio==6.3.0
pillow==12.0.0
numpy==2.3.5
sympy==1.13.3
python-docx==1.2.0
psutil==7.2.1
huggingface-hub==0.36.0
safetensors==0.7.0
tokenizers==0.22.2
sentencepiece==0.2.1
requests==2.32.5
python-multipart==0.0.21
fastapi==0.128.0
uvicorn==0.40.0
